Move ts base related tests to own directory
